requirements.txt
fastapi
uvicorn
langchain-ollama
sentence-transformers
faiss-cpu
numpy
python-multipart
pdfplumber
python-docx
Pillow
httpx
